get the right structure for awkward-cpp
