+++ awkward-cpp/awkward/__init__.py
@@ -1,0 +1,1 @@
+../../awkward/__init__.py
