+++ awkward-cpp/awkward/cpp/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+
+# BSD 3-Clause License; see https://github.com/scikit-hep/awkward-array/blob/master/LICENSE
+
+import numpy
+
+from awkward.cpp.array.jagged import JaggedArrayCpp as JaggedArray
+
+__all__ = ["JaggedArray"]
