+++ awkward-cpp/awkward/cpp/array/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+
+# BSD 3-Clause License; see https://github.com/scikit-hep/awkward-array/blob/master/LICENSE
