+++ awkward-cpp/awkward/cpp/array/jagged.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python
+
+# BSD 3-Clause License; see https://github.com/scikit-hep/awkward-array/blob/master/LICENSE
+
+######################################################################## Numba-accelerated interface
+
+import awkward.array.base
+import awkward.array.jagged
+from .base import CppMethods
+
+from . import _jagged
+
+class JaggedArrayCpp(CppMethods, awkward.array.jagged.JaggedArray):
+    @classmethod
+    def offsets2parents(cls, offsets):
+        _jagged.offsets2parents()
+        return "whatever"
